fix(GUI): kill CLI processes when child writer dies
Turns out that when the parent process gets SIGKILL, the children
process would still remain alive.

In order to mitigate this, we can listen to the `exit` event on the
global `process` object. We could have listened for `SIGKILL`, however
according to the NodeJS documentation:

> SIGKILL cannot have a listener installed, it will unconditionally
> terminate Node.js on all platforms.

Fixes: https://github.com/resin-io/etcher/issues/850
Change-Type: patch
Changelog-Entry: Fix writing process remaining alive after the GUI is closed.
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>